+++ src/class/frontend/popups/class-popup-person.php
@@ -745,12 +745,14 @@ if ( ( isset( $_GET['info'] ) ) && ( $_GET['info'] === 'misc' ) ) {
 		<div class="lumiere_display_flex lumiere_font_em_11 lumiere_align_center">
 			<div class="lumiere_flex_auto lumiere_width_eighty_perc">
 				<div class="identity"><?php echo esc_html( $this->person_name ); ?></div>
-				<div class=""><font size="-1">
+
 				<?php
 
 				# Birth
 				$birthday = $this->person->born() ?? null;
 				if ( isset( $birthday ) && count( $birthday ) !== 0 ) {
+					echo "\n\t\t" . '<div><font size="-1">';
+
 					$birthday_day = isset( $birthday['day'] ) && strlen( $birthday['day'] ) > 0 ? (string) $birthday['day'] . ' ' : __( '(day unknown)', 'lumiere-movies' ) . ' ';
 					$birthday_month = isset( $birthday['month'] ) && strlen( $birthday['month'] ) > 0 ? date_i18n( 'F', $birthday['month'] ) . ' ' : __( '(month unknown)', 'lumiere-movies' ) . ' ';
 					$birthday_year = isset( $birthday['year'] ) && strlen( $birthday['year'] ) > 0 ? (string) $birthday['year'] : __( '(year unknown)', 'lumiere-movies' );
@@ -758,19 +760,20 @@ if ( ( isset( $_GET['info'] ) ) && ( $_GET['info'] === 'misc' ) ) {
 					echo "\n\t\t\t" . '<span class="imdbincluded-subtitle">'
 						. esc_html__( 'Born on', 'lumiere-movies' ) . '</span>'
 						. esc_html( $birthday_day . $birthday_month . $birthday_year );
-				}
 
-				if ( isset( $birthday['place'] ) && strlen( $birthday['place'] ) > 0 ) {
-					/** translators: 'in' like 'Born in' */
-					echo ', ' . esc_html__( 'in', 'lumiere-movies' ) . ' ' . esc_html( $birthday['place'] );
-				}
+					if ( isset( $birthday['place'] ) && strlen( $birthday['place'] ) > 0 ) {
+						/** translators: 'in' like 'Born in' */
+						echo ', ' . esc_html__( 'in', 'lumiere-movies' ) . ' ' . esc_html( $birthday['place'] );
+					}
 
-				echo "\n\t\t" . '</font></div>';
-				echo "\n\t\t" . '<div class=""><font size="-1">';
+					echo "\n\t\t" . '</font></div>';
+				}
 
 				# Death
 				$death = count( $this->person->died() ) > 0 ? $this->person->died() : null;
 				if ( $death !== null ) {
+
+					echo "\n\t\t" . '<div><font size="-1">';
 
 					$death_day = isset( $death['day'] ) && strlen( $death['day'] ) > 0 ? (string) $death['day'] . ' ' : __( '(day unknown)', 'lumiere-movies' ) . ' ';
 					$death_month = isset( $death['month'] ) && strlen( $death['month'] ) > 0 ? date_i18n( 'F', $death['month'] ) . ' ' : __( '(month unknown)', 'lumiere-movies' ) . ' ';
@@ -788,14 +791,15 @@ if ( ( isset( $_GET['info'] ) ) && ( $_GET['info'] === 'misc' ) ) {
 					if ( ( isset( $death['cause'] ) ) && ( strlen( $death['cause'] ) !== 0 ) ) {
 						echo ' (' . esc_html( $death['cause'] . ')' );
 					}
-				}
 
-				echo "\n\t\t" . '</font></div>';
+					echo "\n\t\t" . '</font></div>';
+				}
 
 				echo "\n\t\t" . '<div class="lumiere_display_flex_align_flex_end lumiere_padding_two lumiere_align_left">';
 
-				echo "\n\t\t\t" . '<div><font size="-1">' . wp_kses(
-					$this->link_maker->lumiere_medaillon_bio( $this->person->bio() ) ?? '',
+				$bio = $this->link_maker->lumiere_medaillon_bio( $this->person->bio() );
+				echo is_string( $bio ) && strlen( $bio ) > 0 ? "\n\t\t\t" . '<div><font size="-1">' . wp_kses(
+					$bio,
 					[
 						'span' => [ 'class' => [] ],
 						'a' => [
@@ -807,7 +811,7 @@ if ( ( isset( $_GET['info'] ) ) && ( $_GET['info'] === 'misc' ) ) {
 						'div' => [],
 						'br' => [],
 					]
-				) . '</font></div>';
+				) . '</font></div>' : '';
 				?>
 
 					<!-- star photo -->
@@ -815,14 +819,22 @@ if ( ( isset( $_GET['info'] ) ) && ( $_GET['info'] === 'misc' ) ) {
 
 			// Select pictures: big poster, if not small poster, if not 'no picture'.
 			$photo_url = '';
-			if ( $this->imdb_cache_values['imdbusecache'] === '1' ) { // use IMDBphp only if cache is active
-				$photo_url = $this->person->photo_localurl( false ) !== false ? esc_url( $this->person->photo_localurl( false ) ) : esc_url( $this->person->photo_localurl( true ) ); // create big picture, thumbnail otherwise.
-			}
-			$photo_url_final = strlen( $photo_url ) === 0 ? esc_url( $this->config_class->lumiere_pics_dir . 'no_pics.gif' ) : $photo_url; // take big/thumbnail picture if exists, no_pics otherwise.
+			$photo_big = (string) $this->person->photo_localurl( false );
+			$photo_thumb = (string) $this->person->photo_localurl( true );
 
-			echo "\n\t\t\t\t" . '<a class="highslide_pic_popup" href="' . esc_url( $photo_url_final ) . '">';
+			if ( $this->imdb_cache_values['imdbusecache'] === '1' ) { // use IMDBphp only if cache is active
+				$photo_url = strlen( $photo_big ) > 1 ? esc_url( $photo_big ) : esc_url( $photo_thumb ); // create big picture, thumbnail otherwise.
+			}
+
+			// Picture for a href, takes big/thumbnail picture if exists, no_pics otherwise.
+			$photo_url_href = strlen( $photo_url ) === 0 ? esc_url( $this->config_class->lumiere_pics_dir . 'no_pics.gif' ) : $photo_url; // take big/thumbnail picture if exists, no_pics otherwise.
+
+			// Picture for img: if 1/ thumbnail picture exists, use it, 2/ use no_pics otherwise
+			$photo_url_img = strlen( $photo_thumb ) === 0 ? esc_url( $this->config_class->lumiere_pics_dir . 'no_pics.gif' ) : $photo_thumb;
+
+			echo "\n\t\t\t\t" . '<a class="highslide_pic_popup" href="' . esc_url( $photo_url_href ) . '">';
 			echo "\n\t\t\t\t\t" . '<img loading="lazy" class="imdbincluded-picture" src="'
-				. esc_url( $photo_url_final )
+				. esc_url( $photo_url_img )
 				. '" alt="'
 				. esc_attr( $this->person_name ) . '"';